新加修改密码，修改用户信息接口

+++ src/main/java/com/sim/wifi/upgrade/modules/permissions/service/UserCacheService.java
@@ -15,7 +15,7 @@ public interface UserCacheService {
     /**
      * 删除后台用户缓存
      */
-    void delUser(Long userId);
+    void delUser(Integer userId);
 
     /**
      * 删除后台用户资源列表缓存
